SailBugfix: Correct mvendorid to 32 bits as per RISC-V specification
According to the official RISC-V specification: "The mvendorid CSR is a 32-bit read-only register."

Previously, this register was incorrectly emulated as a 64-bit register, which deviated from the standard. This commit rectifies the issue by implementing mvendorid as a 32-bit register, ensuring compliance with the specification.

+++ src/virt/mod.rs
@@ -143,7 +143,7 @@ pub struct VirtCsr {
     pub mie: usize,
     pub mip: usize,
     pub mtvec: usize,
-    pub mvendorid: usize,
+    pub mvendorid: u32,
     pub marchid: usize,
     pub mimpid: usize,
     pub mcycle: usize,
